+++ Src/Roflcopter.Plugin.Tests/UpdateAspDesignerFiles/AspCodeBehindTestHelper.cs
@@ -1,0 +1,23 @@
+﻿#if RESHARPER
+using System.Linq;
+using JetBrains.TestFramework;
+using JetBrains.Util;
+
+namespace Roflcopter.Plugin.Tests.UpdateAspDesignerFiles
+{
+    internal class AspCodeBehindTestHelper
+    {
+        private readonly BaseTestNoShell _test;
+
+        public AspCodeBehindTestHelper(BaseTest test)
+        {
+            _test = test;
+        }
+
+        private string CodeBehindFileName => _test.TestMethodName + ".aspx.cs";
+        public string DesignerFileName => _test.TestMethodName + ".aspx.designer.cs";
+
+        public string[] ExtendTestFiles(string[] testFiles) => testFiles.Concat(CodeBehindFileName, DesignerFileName).ToArray();
+    }
+}
+#endif