+++ Src/Roflcopter.Plugin.Tests/test/data/UpdateAspDesignerFiles/UpdateAspDesignerFileContextActionTest/MissingDesignerFile.aspx.cs
@@ -1,0 +1,8 @@
+﻿using System.Web.UI;
+
+namespace Roflcopter.Sample.UpdateAspDesignerFiles.UpdateAspDesignerFileContextActionTest
+{
+    public partial class MissingDesignerFile : Page
+    {
+    }
+}