+++ Src/Roflcopter.Plugin.Tests/test/data/UpdateAspDesignerFiles/UpdateAspDesignerFileQuickFixTest/MissingReference.aspx.cs
@@ -1,0 +1,8 @@
+﻿using System.Web.UI;
+
+namespace Roflcopter.Sample.UpdateAspDesignerFiles.UpdateAspDesignerFileQuickFixTest
+{
+    public partial class MissingReference : Page
+    {
+    }
+}